+++ sources/Controller/LandingController.cs
@@ -86,7 +86,7 @@ namespace K2D2.Controller
 
         void LandingUI()
         {
-            burn_before = UI_Tools.FloatSlider("Burn Before", burn_before, 0, 10, "s", "Secure burn before critical time");
+            burn_before = UI_Tools.FloatSlider(burn_before, "Burn Before", 0, 10, "s", "Secure burn before critical time");
         }
 
         void WarpUI()
@@ -97,13 +97,12 @@ namespace K2D2.Controller
             {
                 WarpToSettings.onGUI();
 
-
                 if (Settings.debug_mode)
                 {
                     UI_Tools.Title("// Rotation Warp");
                     rotation_warp_duration = UI_Fields.IntField("rotation_warp_duration", "Rot. Warp Duration", rotation_warp_duration, 0, int.MaxValue,
                     "During Rotation Warp, Attitude is checked");
-                    max_rotation = UI_Tools.FloatSlider("Safe Warp Rotation", max_rotation, 0, 90, "°", "Max angle (stop warp when reached)");
+                    max_rotation = UI_Tools.FloatSlider(max_rotation, "Safe Warp Rotation",0, 90, "°", "Max angle (stop warp when reached)");
                 }
                 else
                 {
@@ -116,24 +115,22 @@ namespace K2D2.Controller
         void TouchDown_UI()
         {
             UI_Tools.Title("// Touch Down");
-            start_touchdown_altitude = UI_Tools.FloatSlider("Start TouchDown Altitude", start_touchdown_altitude, 100, 5000, "m", "Altitude for starting Touch-Down Phase");
+            start_touchdown_altitude = UI_Tools.FloatSlider(start_touchdown_altitude, "Start TouchDown Altitude", 100, 5000, "m", "Altitude for starting Touch-Down Phase");
 
-            touch_down_ratio = UI_Tools.FloatSlider("Altitude/speed ratio", touch_down_ratio, 0.5f, 3, "", "Speed is based on altitude");
+            touch_down_ratio = UI_Tools.FloatSlider(touch_down_ratio, "Altitude/speed ratio", 0.5f, 3, "", "Speed is based on altitude");
             UI_Tools.Right_Left_Text("Safe", "Danger");
 
-            touch_down_speed = UI_Tools.FloatSlider("Touch-Down speed", touch_down_speed, 0.1f, 10, "m/s", "Speed when touching ground");
+            touch_down_speed = UI_Tools.FloatSlider(touch_down_speed, "Touch-Down speed", 0.1f, 10, "m/s", "Speed when touching ground");
 
         }
-
-
 
         public void settings_UI()
         {
             if (accordion.Count == 0)
             {
-                accordion.addChapter("// Landing Settings", LandingUI);
-                accordion.addChapter("// Warp", WarpUI);
-                accordion.addChapter("// Touch Down", TouchDown_UI);
+                accordion.addChapter("Landing Settings", LandingUI);
+                accordion.addChapter("Warp", WarpUI);
+                accordion.addChapter("Touch Down", TouchDown_UI);
                 accordion.singleChapter = true;
             }
 
@@ -216,7 +213,7 @@ namespace K2D2.Controller
                     else
                     {
                         current_executor.setController(warp_to);
-                        warp_to.Start_UT(startSafeWarp_UT);
+                        warp_to.Start_Retrograde(startSafeWarp_UT);
                     }
                     break;
                 case Mode.RotationWarp:
@@ -226,7 +223,7 @@ namespace K2D2.Controller
                     else
                     {
                         current_executor.setController(warp_to);
-                        warp_to.Start_UT(startBurn_UT, true);
+                        warp_to.Start_Retrograde(startBurn_UT, true);
                     }
                     break;
                 case Mode.Waiting: