+++ sources/Controller/SimpleManeuverController.cs
@@ -52,6 +52,13 @@ namespace K2D2.Controller
 
         public override void onGUI()
         {
+            if (K2D2_Plugin.Instance.settings_visible)
+            {
+                Settings.onGUI();
+                return;
+            }
+
+
             if (_maneuverProvider == null)
             {
                 return;