+++ sources/InfosPages/SASInfos.cs
@@ -10,6 +10,12 @@ namespace K2D2.InfosPages
     {
         public override void onGUI()
         {
+            if (K2D2_Plugin.Instance.settings_visible)
+            {
+                Settings.onGUI();
+                return;
+            }
+
             var current_vessel = K2D2_Plugin.Instance.current_vessel.VesselComponent;
 
             var sas = current_vessel.Autopilot.SAS;