+++ sources/InfosPages/VesselInfos.cs
@@ -10,6 +10,12 @@ namespace K2D2.InfosPages
     {
         public override void onGUI()
         {
+            if (K2D2_Plugin.Instance.settings_visible)
+            {
+                Settings.onGUI();
+                return;
+            }
+
             var vehicle = K2D2_Plugin.Instance.current_vessel.VesselVehicle;
             var vessel_component = K2D2_Plugin.Instance.current_vessel.VesselComponent;