+++ sources/Tools/GeneralTools.cs
@@ -51,7 +51,6 @@ namespace K2D2
             return dt;
         }
 
-
         public static Guid createGuid()
         {
             return Guid.NewGuid();